Former-commit-id: 8f9c9f8239858d5b056b23aea9f1c68d21b5f4ef

+++ SpyFamily/src/main/java/com/ssafy/spyfamily/chat/model/ChatRoom.java
@@ -28,5 +28,5 @@ public class ChatRoom {
         return chatRoom;
     }
     @ColumnDefault("0")
-    private Long loginUserCount;
+    private Integer loginUserCount;
 }
